Setting up armor regen and dodge damage over time.

+++ src/main/java/com/rubbishman/rubbishcombat/stages/DodgeStage.java
@@ -1,11 +1,10 @@
 package com.rubbishman.rubbishcombat.stages;
 
-import com.rubbishman.rubbishRedux.experimental.actionTrack.stage.StageProcessor;
-import com.rubbishman.rubbishRedux.experimental.actionTrack.stage.StageWrapResult;
-import com.rubbishman.rubbishRedux.experimental.actionTrack.stage.StageWrappedAction;
 import com.rubbishman.rubbishRedux.external.operational.store.ObjectStore;
+import com.rubbishman.rubbishRedux.external.setup_extra.actionTrack.stage.StageProcessor;
+import com.rubbishman.rubbishRedux.external.setup_extra.actionTrack.stage.StageWrapResult;
+import com.rubbishman.rubbishRedux.external.setup_extra.actionTrack.stage.StageWrappedAction;
 import com.rubbishman.rubbishcombat.actions.Damage;
-import com.rubbishman.rubbishcombat.actions.internalcombat.ArmorDamage;
 import com.rubbishman.rubbishcombat.actions.internalcombat.DodgeDamage;
 import com.rubbishman.rubbishcombat.state.CombatEntity;
 
@@ -16,7 +15,7 @@ public class DodgeStage implements StageProcessor {
 
         CombatEntity combatEntity = state.getObject(damage.target);
 
-        int dodgeTaken = Math.min(damage.damage, combatEntity.dodge.factor);
+        int dodgeTaken = (int)Math.min(damage.damage, combatEntity.dodge.factor);
         int remainingDamageTaken = Math.max(damage.damage - dodgeTaken, 0);
 
         return new StageWrapResult(
