+++ src/main/java/com/rubbishman/rubbishcombat/state/attribute/DodgeAttribute.java
@@ -1,12 +1,12 @@
 package com.rubbishman.rubbishcombat.state.attribute;
 
 public class DodgeAttribute {
-    public final int current;
-    public final int factor;
-    public final int period;
-    public final int periodFactor;
+    public final long current;
+    public final long factor;
+    public final long period;
+    public final long periodFactor;
 
-    public DodgeAttribute(int current, int factor, int period, int periodFactor) {
+    public DodgeAttribute(long current, long factor, long period, long periodFactor) {
         this.current = current;
         this.factor = factor;
         this.period = period;
